# [SPARK-37899][SQL] Support converting inner join to left semi join 