@wangyum 2022-07-20
A production case.
Before this PR | After this PR
-- | --
![image](https://user-images.githubusercontent.com/5399861/179954626-b3d74829-ab86-4831-a0a0-5b550142741d.png) | ![image](https://user-images.githubusercontent.com/5399861/179954700-42854a07-ca69-4a54-ae65-9d7825c4b417.png)