@wangyum 2022-07-23
Close it because this idea needs statistics. We should not convert to a left semi-join if the left side can broadcast and the right side cannot.

Another idea is to push this join through join . But I think it's too special.